Create README.md
Greetings. My name is Konstantinos Grousouzakos and I am aspiring to become a Data Analyst/Scientist. My educational background consists of a Bachelor's in Mechanical Engineering and a Master's in Industrial Engineering. My passion for statistics and data driven decision making has led me to choose the profession of Data Analyst/Scientist as the one that fits best to my temperament, personality and natural skills. In this portfolio, you will find projects that I did on my own, such as EDA, predictive models, classification and clustering models. The portfolio will be regularly updated. Any feedback is welcome.
Here is the hyperlink to my Linkedin profile : www.linkedin.com/in/konstantinos-grousouzakos-data-analyst